Allow for referencing eslint and prettier configs as packages pla-92
