+++ eslint-config-ofmt/eslint.tailwind.cjs
@@ -1,4 +1,3 @@
-/* eslint-disable max-len, max-lines -- disabling file length limitation for compactness. */
 const groups = [
   {
     members: [
