+++ eslint-config-ofmt/package.json
@@ -1,13 +1,8 @@
 {
-  "name": "@ottofeller/ofmt",
-  "version": "1.0.3",
-  "description": "Autoformatter built for OttoFeller projects",
-  "bin": {
-    "ofmt": "./bin/ofmt.js",
-    "olint": "./bin/olint.js"
-  },
+  "name": "@ottofeller/eslint-config-ofmt",
+  "version": "1.1.0",
+  "description": "ESLint config files for @ottofeller/ofmt",
   "dependencies": {
-    "@ottofeller/eslint-plugin-ottofeller": "0.1.3",
     "@typescript-eslint/eslint-plugin": "5.10.2",
     "@typescript-eslint/parser": "5.10.2",
     "eslint": ">= 8",
@@ -18,6 +13,7 @@
     "meow": "10.1.2"
   },
   "devDependencies": {
+    "@ottofeller/eslint-plugin-ottofeller": "0.1.3",
     "@types/node": "17.0.23",
     "eslint-plugin-eslint-comments": "3.2.0",
     "typescript": "4.5.5"
@@ -37,11 +33,11 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/ottofeller/eslint-config-ottofeller.git"
+    "url": "git+https://github.com/ottofeller/ofmt.git"
   },
   "bugs": {
-    "url": "https://github.com/ottofeller/eslint-config-ottofeller/issues"
+    "url": "https://github.com/ottofeller/ofmt/issues"
   },
-  "homepage": "https://github.com/ottofeller/eslint-config-ottofeller#readme",
+  "homepage": "https://github.com/ottofeller/ofmt#readme",
   "type": "module"
 }
